requirements.txt
Pillow>=10.0.0
imagetext-py>=1.2.0
phomemo-printer>=0.2.0
jsonschema>=4.19.0
python-dotenv>=1.0.0
mcp>=0.1.0
pydantic>=2.6.0
